Set persist.audio.fluence.speaker to false
There's been a lot of complaints of distortion/echo-ing when calling on
speaker for a very long time.
The echo-ing and distortion renders calling on speaker almost completely
useless.
Having persist.audio.fluence.speaker set to true seems to be causing the
issues we're having. For now we should set persist.audio.fluence.speaker
to false to be able to use calliing on speaker properly, and in the
meantime sort out what's going wrong.
Currently I'm working through the audio HAL code to figure out how
the persist.audio.fluence.speaker code is handled and what can be done
to fix it for the G2.

Change-Id: I5cfcfb6b7ed3261143978528474ed85d1b161953